+++ bc/location/blocks.py
@@ -1,0 +1,6 @@
+from bc.service_directory.blocks import DirectoryServicesBlock
+from bc.utils.blocks import StoryBlock
+
+
+class LocationPageStoryBlock(StoryBlock):
+    directory_services = DirectoryServicesBlock()