+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Compare: <https://git.torchbox.com/buckinghamshire-council/bc/compare/77.0...78.0>
 
 - [Directory search block](https://trello.com/c/09eNT1iH/6-fis-directory-search-widget-fe-and-be) ([Merge request #712](https://git.torchbox.com/buckinghamshire-council/bc/-/merge_requests/712))
+- [Directory services block](https://trello.com/c/CcEf4OYn) ([Merge request #727](https://git.torchbox.com/buckinghamshire-council/bc/-/merge_requests/727))
 
 ## 77.0 (2024-07-29)